requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncssh==2.14.2
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
python-telegram-bot==20.7
jinja2==3.1.3
python-multipart==0.0.6
cryptography==42.0.0
